+++ src/components/Toc.js
@@ -1,0 +1,17 @@
+import React, { Component } from 'react'; //React는 필수.
+
+class Toc extends Component {
+    render() {
+      return (
+        <nav>
+          <ul>
+            <li><a href="1.html">HTML</a></li>
+            <li><a href="2.html">CSS</a></li>
+            <li><a href="3.html">JavaScript</a></li>
+          </ul>
+        </nav>
+      );
+    }
+}
+
+export default Toc;
